permission-all

+++ app/Http/Controllers/GroupController.php
@@ -14,20 +14,24 @@ class GroupController extends Controller
 {
    
     public function index()
-    {   $groups = Group::search()->latest()->paginate(5);
+    {   
+        $this->authorize('viewAny', Group::class);
+        $groups = Group::search()->latest()->paginate(5);
         return view('admin.groups.index', compact('groups'));
     }
 
   
     public function create()
     {   
+        $this->authorize('Group_create', Group::class);
         $parent_roles = Role::where('group_key', 0)->get();
         return view('admin.groups.create',compact('parent_roles'));
     }
 
   
     public function store(StoreGroupRequest $request)
-    {
+    {   
+        $this->authorize('Group_create', Group::class);
         try {
             $role = Group::create([
                 'name' => $request->name,
@@ -51,7 +55,7 @@ class GroupController extends Controller
     public function edit(string $id)
     {
         try {
-
+            $this->authorize('Group_update', Group::class);
             $group = Group::find($id);
             $roles_checked = $group->roles;
             $parent_roles = Role::where('group_key', 0)->get();
@@ -71,6 +75,7 @@ class GroupController extends Controller
    
     public function update(UpdateGroupRequest $request,$id)
     {
+        $this->authorize('Group_update', Group::class);
         try {
             $group = Group::find($id);
             $group->update([
@@ -86,7 +91,8 @@ class GroupController extends Controller
     }
 
     public function destroy(string $id)
-    {
+    {   
+        $this->authorize('Group_delete', Group::class);
         try {
             $group = Group::find($id);
             DB::table('group_roles')->where('group_id', '=', $group->id)->delete();
